requirements.txt
# Core utilities
python-dotenv==1.0.1
tqdm==4.66.5

# Data & Math
numpy==1.26.4
pandas==2.2.3
scipy==1.16.3
scikit-learn==1.5.2
sympy==1.14.0
mpmath==1.3.0
tbb==2021.13.1
intel-openmp==2021.4.0

# ML / NLP
torch==2.3.1
transformers==4.43.3
sentence-transformers==2.7.0
huggingface-hub==0.36.0
tokenizers==0.19.1
tiktoken==0.12.0
safetensors==0.6.2

# LangChain ecosystem (stable range)
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.4
langchain-openai==0.2.6
langchain-text-splitters==0.3.0
langsmith==0.1.147

# Database & Cloud
supabase==2.6.0
postgrest==0.16.10
storage3==0.7.7
gotrue==2.12.4
supafunc==0.5.1
psycopg2-binary==2.9.9

# Vector DB
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7

# PDF Processing
PyMuPDF==1.23.8
PyMuPDFb==1.23.7

# API Frameworks
fastapi==0.112.4
starlette==0.37.2
uvicorn==0.25.0
python-multipart==0.0.9

# Frontend & Visualization
streamlit==1.40.2
streamlit-lottie==0.0.5
altair==5.5.0
pydeck==0.9.1
rich==13.9.4

# Chainlit App
chainlit==1.2.0

# Network & Auth
httpx==0.27.2
httpcore==1.0.9
requests==2.32.5
urllib3==2.5.0
PyJWT==2.10.1
cryptography==46.0.3
bcrypt==5.0.0
aiohttp==3.13.2
aiofiles==23.2.1
asyncer==0.0.7
anyio==4.11.0

# Logging / Tracing / Telemetry
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
uptrace==1.37.0

# Misc utilities
python-dateutil==2.9.0.post0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.8.1
dataclasses-json==0.6.7
typing-extensions==4.11.0
typing-inspect==0.9.0
attrs==25.4.0
tenacity==9.1.2
toml==0.10.2
tomli==2.3.0
packaging==23.2
jinja2==3.1.6
colorama==0.4.6
markdown-it-py==4.0.0
mdurl==0.1.2
deprecation==2.1.0
jsonschema==4.25.1
jsonpatch==1.33
zipp==3.23.0
idna==3.11
charset-normalizer==3.4.4
six==1.17.0
watchdog==6.0.0
watchfiles==0.20.0
websockets==12.0
propcache==0.4.1
blinker==1.9.0
filetype==1.2.0
chevron==0.14.0
realtime==1.0.5
strenum==0.4.15
yarl==1.22.0
